+++ PriceCalculatorKata/ExpenseCalculator.cs
@@ -2,10 +2,10 @@
 
 public static class ExpenseCalculator
 {
-    public static float CalculateExpenses(Product product)
+    public static decimal CalculateExpenses(Product product)
     {
         if (!product.HasExpenses()) return 0;
-        float totalExpenses = 0;
+        decimal totalExpenses = 0;
         foreach (var expense in product.Expenses)
         {
             if (expense.ValueType.Equals(Constants.ValueType.Value))
